+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed `micropip.add_mock_package` to work with Pyodide>=0.23.0
+  [#66](https://github.com/pyodide/micropip/pull/66)
+
 ## [0.3.0] - 2023/03/29
 
 ### Added