+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `verbose` parameter to micropip.install and micropip.uninstall
   [#60](https://github.com/pyodide/micropip/pull/60)
 
+### Fixed
+
+- Fixed `micropip.add_mock_package` to work with Pyodide>=0.23.0
+  [#66](https://github.com/pyodide/micropip/pull/66)
+
 ## [0.3.0] - 2023/03/29
 
 ### Added