+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.4.0] - 2023/07/25
+
 ### Added
 
 - Added `verbose` parameter to micropip.install and micropip.uninstall